Added Api controller

+++ LinkShortenerWebApi/Repositories/ILinksRepository.cs
@@ -7,10 +7,10 @@ namespace LinkShortenerWebApi.Repositories
  public interface ILinksRepository
     {
         (IEnumerable<Link>, int) Get(string search, int skip);
-        Link Get(int Id);
+        Link Get(long Id);
         Link Create(Link link);
         Link Update(Link link);
-        void Delete(int id);
+        void Delete(long id);
         string GetUrlForHash(string hash);
     }
 }
